requirements.txt
django 
django-cors-headers
djangorestframework 
pillow 

